440: Split elections controller into separate files for each use case
