+++ apps/backend/src/routes/elections.routes.ts
@@ -16,16 +16,14 @@ import { deleteCandidate } from '../controllers/candidates/deleteCandidate.uc.js
 import { getCandidate } from '../controllers/candidates/getCandidate.uc.js';
 import { getCandidates } from '../controllers/candidates/getCandidates.uc.js';
 import { updateCandidate } from '../controllers/candidates/updateCandidate.uc.js';
-import {
-  createElection,
-  deleteElection,
-  freezeElection,
-  getElection,
-  getElections,
-  getFreezableElection,
-  unfreezeElection,
-  updateElection,
-} from '../controllers/elections.controllers.js';
+import { createElection } from '../controllers/elections/createElection.uc.js';
+import { deleteElection } from '../controllers/elections/deleteElection.uc.js';
+import { freezeElection } from '../controllers/elections/freezeElection.uc.js';
+import { getElection } from '../controllers/elections/getElection.uc.js';
+import { getElections } from '../controllers/elections/getElections.uc.js';
+import { getFreezableElection } from '../controllers/elections/getFreezableElection.uc.js';
+import { unfreezeElection } from '../controllers/elections/unfreezeElection.uc.js';
+import { updateElection } from '../controllers/elections/updateElection.uc.js';
 import { acceptBodyCheck } from '../middlewares/acceptBodyCheck.js';
 import { acceptHeaderCheck } from '../middlewares/acceptHeaderCheck.js';
 import { defaultBallotPaperChecks } from '../middlewares/pathParamChecks/ballotPaperChecks.js';
